requirements.txt
# Core dependencies for Richmond Tech Community demo
boto3>=1.34.0               # AWS SDK for Python
botocore>=1.34.0            # AWS core functionality

# Strands SDK and MCP
strands-agents>=1.1.0         # Official Strands AI agent framework
strands-agents-tools>=0.1.7   # Pre-built tools for Strands agents
mcp[cli]>=1.4.0              # Model Context Protocol with CLI support

# Claude/Anthropic API
anthropic>=0.8.0            # Anthropic Claude API client

# Data processing
python-dateutil>=2.8.2     # Date/time parsing
typing-extensions>=4.8.0   # Enhanced type hints

# Logging and utilities
structlog>=23.1.0           # Structured logging
click>=8.1.0                # CLI interface
pydantic>=2.5.0             # Data validation
python-dotenv>=1.0.0        # Environment variable management

# Development dependencies
pytest>=7.4.0               # Testing framework
pytest-asyncio>=0.21.0      # Async testing
pytest-mock>=3.11.0         # Mocking for tests
black>=23.0.0               # Code formatting
flake8>=6.0.0               # Linting
mypy>=1.7.0                 # Type checking

# AWS Lambda dependencies (for deployment)
aws-lambda-powertools>=2.25.0  # Lambda utilities
mangum>=0.17.0              # ASGI adapter for Lambda

# JSON and data handling
ujson>=5.8.0                # Fast JSON parsing
orjson>=3.9.0               # Fast JSON serialization

# HTTP and CLI interface
httpx>=0.25.0               # Modern HTTP client
rich>=13.0.0                # Rich terminal output